This primarily resolves #338, so Eta now has basic concurrency. 

Details:
- If you run `forkIO` more than say N times, where N is the value of the `+RTS -N[N]` flag you send when you run the program, there is no guarantee that your forked threads' tasks will be given fairness in CPU time. If you use it less than N times, you are guaranteed that the tasks will run separately. See #358.
- The threaded runtime is now default because no one writes useful non-concurrent applications anymore.
- MVar is implemented to work with the new `forkIO` change. STM primitives need to updated, see #360.
- A minor data structure change was made making the Eta stack use a `LinkedList` instead of a `Stack`. The result was a 20-30% increase in performance across the board, with an increase in GC allocation rate by up to 2x in some cases. This can be addressed with an intrusive list implementation, but for now the JVM GC is good enough to handle it.

[Before](https://circleci.com/gh/typelead/eta-benchmarks/45#tests/containers/0) and [After](https://circleci.com/gh/typelead/eta-benchmarks/46#build-timing/containers/0) (Click on `./eta-benchmarks.sh`)

Some sample entries here:

Benchmark | Before (ms) | After (ms)
--- | --- | ---
`imaginary/exp3_8` | 1018 | 718
`imaginary/wheel-sieve1` | 3585 | 2522
`spectral/lambda` | 9341 | 6542