@ashishnegi 2017-05-08
About the usage of `linked list vs vector` stacks, this link might be useful : http://agis.io/2014/03/25/contiguous-stacks-in-go.html 

The concept is called `stack thrashing`. 

Quoting page :
```
In such cases where the stack boundary happens to fall in a tight loop,
the overhead of creating and destroying segments repeatedly becomes significant. 
This is called the “hot split” problem inside the Go community. 
Rust folks call it “stack thrashing”.
```

So, benchmarking on this pathological case will be interesting.