@rahulmutt 2017-05-08
Thanks for the share. Got me thinking about ways to optimise it even more! 

In the case of the Eta stack, alternate addition/removal doesn't seem to happen in practice as it does in Go, simply because most tail-recursive calls are optimised into loops hence stay within the same method and keep the stack relatively constant. In most cases you'll have a long string of additions followed by strings of removals and that doesn't happen nearly as often as it does in Go. The 30% increase was actually because of speeding up removes!

For the stack operations to be fast, we need O(1) insertions to the top *and* O(1) removals at arbitrary points (there are cases where just removing from the top is not enough). LinkedList clearly wins with such requirements, and Vector is slow since it does array copies on every removal.